File.php now creates or deletes file 'public' on save() rather than too early

+++ core/File.php
@@ -62,7 +62,7 @@ class File extends DBRow {
 				error_log ("Key $key=$value not recognized in getLink");
 			}
 		}
-		$args = implode ("&amp;", $results);
+		$args = implode ("&", $results);
 		if ($args) $args = "?$args";
 		return "/im$path$args";
 	}
@@ -81,6 +81,8 @@ class File extends DBRow {
 	private function getLocalFilename() {
 		// Replaces the extension with one matching the mime type, if available
 		$filename = $this->getFilename();
+		$filename = preg_replace('/ /', '_', $filename);
+		$filename = preg_replace('/[^0-9a-zA-Z._]/', '', $filename);
 		if (!$filename) return "file.tmp";
 		$pos = strrpos ($filename, '.');
 		if ($pos === false) return "$filename.tmp";
@@ -90,13 +92,15 @@ class File extends DBRow {
 		if (!$ext) $ext = "tmp";
 		return substr($filename, 0, $pos) . ".$ext";
 	}
+
+	function getFullPath(){return $this->getDirectory() . $this->getLocalFilename();}
 	
-	function insertNotWeb($tempurl, $type, $filename) {
+	private function _insert($tempurl, $type, $filename, $copy = true) {
 		if (!is_readable ($tempurl)) {
-			error_log ("Upload of file ".$tmp." failed.");
+			error_log ("Upload of file ".$tmpurl." failed.");
 			return false;
 		}
-		
+		// Assume the file is uploaded (for now) and set up the file object; need to save() to get a new id.
 		$this->setType ($type);
 		$this->setFilename ($filename);
 		if (!$this->getId()) $this->save();
@@ -112,38 +116,20 @@ class File extends DBRow {
 		mkdir($dir,0777,true);
 		umask($oldmask);
 
-		if (!copy ($tempurl, $dir.$file)) {
-			error_log ("Move of file $tempurl to $dir.$file failed.");
+		$method = $copy ? 'copy' : 'move_uploaded_file';
+		$Method = $copy ? 'Copy' : 'Move';
+		if (!$method ($tempurl, $dir.$file)) {
+			error_log ("$Method of file $tempurl to $dir$file failed.");
 			return false;
 		}
 	}
 	
-	function insert($data) {
+	function insert($data, $type = null, $filename = null) {
+		// $data is either an array OR a quickform result OR a URL to be copied;
+		// if the latter, proved $type and $filename
 		if ($data instanceof HTML_QuickForm_file) {$data = $data->getValue();}
-
-		// First check if there is a file available for upload
-		$tmp = $data['tmp_name'];
-		if (!is_readable ($data['tmp_name'])) {
-			error_log ("Upload of file $tmp failed.");
-			return false;
-		}
-
-		// Assume the file is uploaded (for now) and set up the file object; need to save() to get a new id.
-
-		$this->setType ($data['type']);
-		$this->setFilename ($data['name']);
-		if (!$this->getId()) $this->save();
-		if (!$this->getId()) {
-			error_log ("Save failed in File.php");
-			return false;
-		}
-		$dir = $this->getDirectory();
-		$file = $this->getLocalFilename();
-		mkdir($dir,0777,true);
-		if (!move_uploaded_file ($data['tmp_name'], $dir.$file)) {
-			error_log ("Upload of file $tmp to $loc failed.");
-			return false;
-		}
+		if (is_array ($data)) return $this->_insert($data['tmp_name'], $data['type'], $data['name'], false);
+		else                  return $this->_insert($data, $type, $filename, true);
 	}
 
 	function getAddEditFormHook($form) {
@@ -152,13 +138,13 @@ class File extends DBRow {
 
 	function setPublic() {$this->setPermission('public');}
 	function setPrivate() {$this->setPermission('private');}
-	function setPermission($perm) {
-		$this->permission = $perm;
+	function &save(&$notification = null) {
+		parent::save($notification);
+		if (!$this->getId()) return; // Just a security precaution; this case should not happen.
 		$file = $this->getDirectory() . 'public';
 		if ($perm == 'public') touch($file);
 		else unlink($file);
 	}
-	
 	function getAddEditFormSaveHook($form) {
 		$el = $form->getElement('upload_file');
 		if ($el->isUploadedFile()) {
